adjust statistics

+++ MyStore.Application/Services/Sizes/ISizeService.cs
@@ -7,7 +7,7 @@ namespace MyStore.Application.Services.Sizes
     {
         Task<IEnumerable<SizeDTO>> GetSizesAsync();
         Task<SizeDTO> AddSizeAsync(string name);
-        Task<SizeDTO> UpdateSizeAsync(int id, string name);
-        Task DeleteSizeAsync(int id);
+        Task<SizeDTO> UpdateSizeAsync(long id, string name);
+        Task DeleteSizeAsync(long id);
     }
 }
